requirements.txt
requests~=2.23.0
python-digitalocean~=1.15.0
prometheus-client~=0.7.1
